Redid swerve drive command
Made swervedrive command universal, and now pilotSwerve is in the PilotCommand class. Lets us make a lot more drive commands based on sending x, y, and r speeds directly to swerve. Can handle the math in Pilot classes which makes sense anyway. Also renamed snap to aim, so it can be used for aiming from a limelight or anything else that gives us an angle to turn to.